Revert "Use the same system cert path for mac as we do windows (#10595)"
This reverts commit 61d07c18c1de7d6a51a9a2d671b963b25893b38d.
